+++ packages/elvis-assets-icons/config/icons.config.js
@@ -480,10 +480,14 @@ module.exports = [
   { name: 'phone_image', terms: ['mobile', 'picture', 'content'], thirdparty: { name: ['streamline'] } },
   { name: 'pdf_document', terms: ['adobe', 'file'], thirdparty: { name: ['streamline'] } },
   { name: 'pie_chart', terms: [], thirdparty: { name: ['streamline'] } },
-  { name: 'piggy_bank-color', terms: [], thirdparty: { name: ['streamline'] } },
+  {
+    name: 'piggy_bank-color',
+    terms: ['save', 'saving', 'heart', 'cost', 'money', 'expenses', 'green', 'safe', 'reduce'],
+    thirdparty: { name: ['streamline'] },
+  },
   {
     name: 'pin',
-    terms: ['save', 'saving', 'heart', 'cost', 'money', 'expenses', 'green', 'safe', 'reduce'],
+    terms: [],
     thirdparty: { name: ['streamline'] },
   },
   {